Stop the menu losing the top half of every handle it is given
Brian's AutoLoot took nothing from a chest. He had already hit it once, filed
inc-802, and hit it again on a new save with a second, different chest -- which
is what unpicked the recorded diagnosis. That issue's remaining hypothesis was
Item::Owner() walking up through a chest that had kept the player as its
parent. A Parent/Owner probe in src/Dump.cpp read the chain and refuted it: the
oak chest is Parent=0 and every item in it is Owner=0, the same as the ground
items that worked.

The handle is destroyed on its way INTO the menu, not read wrongly on the way
out. inc/Api.h declares the script side of every engine call and the resource
compiler turns those declarations into the virtual machine's dispatcher. The
declaration for LOption -- the call every script uses to add one line to a menu
-- said its value parameter was an int16, so lib/dispatch.h read

    T1->LOption(GETSTR(STACK(1)), (int16)STACK(2), GETSTR(STACK(3)));

That value is the caller's own cookie, and AutoLoot puts the item's handle in
it (lib/alchemy.irh:1241). An hObj is a signed 32-bit handle. A probe on
TextTerm::FirstSelected read the loss directly: handle 44177 came back out of
the menu as -21359, which is 44177 as a signed 16-bit number. No object answers
to that, AutoLoot's own first guard is !isValidHandle, and the macro walked
past every marked item without a word and said "finished looting".

Handles are handed out in order from 128, so the first 32639 objects of a game
are untouched and the defect switches on partway through a character's life.
That is why Brian's first chest looted and his second did not, and why the
earlier discriminating experiment pointed at "in a chest": the ground robes it
compared against were old inventory objects carrying a low handle. The same
declaration also narrows AutoDrop (lib/alchemy.irh:1062) and every resource id
a script puts in a menu -- lib/abilities.irh:98-237, the undead-form menus, put
ids past sixteen million through it.

TextTerm::LOption itself takes an int32 and Option::Val is an int32, so only
the declaration was ever narrow. Widened it and regenerated lib/dispatch.h;
one line of that file changed and the rest came back byte-identical, which is
also the evidence that regenerating it is safe. LMenu's return value and
FirstSelected/NextSelected were checked and are not cast; the stati family's
int16 parameters match the engine's own int16 fields, so they are not the same
defect wearing a different hat.

tools/check_menu_value.sh is the check. No scripted run is long enough to reach
a wide handle -- a fresh character's gear carries handle 139 -- so
INCURSION_HANDLE_BASE, read once in Registry::Registry, starts the counter
wherever the caller asks. The check sets it to 40000 and drives AutoDrop, whose
menu round-trip is the same one. Confirmed red on a build differing in nothing
but that cast ("AutoDrop marked an item and dropped nothing") and green with
it. The same script at the ordinary base of 128 passes on both builds, which is
the control: without it, an AutoDrop broken for some other reason would fail
this check and be read as the truncation returning.

Verified on a sandboxed copy of Brian's own save, never on save/ itself: the
keys that printed "You stop AutoLoot (finished looting)" and took nothing now
print "? spellbook [Conjurations and Summonings] taken.", and four marked items
come across in four ticks. The 40-seed gate is unchanged against its baseline
-- 40 of 40 reaching a map, 3 error lines, 2 audit findings, 24 died-or-stuck,
7 threat-frozen, all equal -- and check_headless, check_dump_save,
check_store_scroll, check_target_order, check_sacrifice, check_stair_cycle,
check_quiet_lookup and check_sharp_senses all pass. check_app and check_package
were not run; they build the signed app.

Upstream, and marked as such. hObj is 32 bits on Win32 too (signed long there,
signed int here), the cast truncates identically on the upstream compiler, and
inc/Api.h has not been touched since rmtew's own commits of 2014-2015. Not
sent.

+++ lib/dispatch.h
@@ -2095,7 +2095,7 @@ void VMachine::CallMemberFunc(int16 funcid, hObj h, int8 n) {
         return;
     case 412:
         ASSERT(REGS(63) >= 3)
-        T1->LOption(GETSTR(STACK(1)), (int16)STACK(2), GETSTR(STACK(3)));
+        T1->LOption(GETSTR(STACK(1)), STACK(2), GETSTR(STACK(3)));
         return;
     case 413:
         ASSERT(REGS(63) >= 0)
